+++ extend/bajo/hook/dobo.sumba-team-guard@after-action.js
@@ -1,0 +1,6 @@
+async function afterAction (action, ...args) {
+  if (!['createRecord', 'updateRecord', 'removeRecord'].includes(action)) return
+  await this.getTeamGuards(true)
+}
+
+export default afterAction