+++ extend/waibuDb/schema/route-guard.js
@@ -1,0 +1,15 @@
+async function routeGuard () {
+  return {
+    common: {
+      widget: {
+        teams: {
+          attr: {
+            refUrl: 'waibuAdmin:/{prefix}/team/details?id={id}'
+          }
+        }
+      }
+    }
+  }
+}
+
+export default routeGuard